requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
exceptiongroup==1.2.2
faiss-gpu==1.7.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-ollama==0.1.3
langchain-text-splitters==0.2.4
langsmith==0.1.147
MarkupSafe==2.1.5
marshmallow==3.22.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
ollama==0.4.4
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pillow==10.4.0
pkg_resources==0.0.0
propcache==0.2.0
psutil==6.1.0
PyAutoGUI==0.9.54
pydantic==2.10.3
pydantic_core==2.27.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
shiboken6==6.6.3.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
